Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> Is this update expected to cause rendering changes? I have a few PDFs that no longer render like before. The differences are very minor and not visible to the eye, but it is no longer pixel perfect.

Yes and no.

As-is, there seems to be differences when the CPU supports AVX2.
I haven't been able to investigate to see if these are deliberate or not.

However, even though this patch landed the DEPS roll did not. Does maybe PDFium depend on this directory without a DEPS roll being involved? If so, that would mean PDFium and Chromium are currently on different libjpeg_turbo versions. I would imagine that isn't the case...?

Patch-set: 13