Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> LGTM.

Thanks for the review - but I'm not seeing a button to send this to the CQ? Are owners the only ones with the right to CQ patches on third party repos? 

> If we want, we could have follow-up patches to explicitly add each SIMD type's alignment restriction and then error if it is a new, unknown type. That would prevent someone missing it later on. For example, imagine AVX30 requires 128-byte alignment. Having the default be "I don't recognize this SIMD type. You have to tell me its alignment here." would be nice.

Yes this would be nicer than blanket checks. Having architecture specific values would be beneficial in terms of memory usage too. At the moment everything uses 32-byte alignment which means on Arm - only requiring 16-byte alignment - we're wasting a minimum of 16 bytes in every libjpeg-turbo memory allocation. (Every allocation request is rounded up to the next multiple of ALIGN_SIZE, 2 * ALIGN_SIZE in the case of sample buffers.)

I'll add this to my list of things to do :)

Patch-set: 1
Reviewer: Gerrit User 1247893 <1247893@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>