Update patch set 1
Patch Set 1:

(9 comments)

> I have reviewed a large chunk but haven't yet reviewed the general case.
> I want to send this out so I can bike home before the rain starts. :)

Thanks for taking a look :)

> I think we have some extra cases to consider. Specifically, top left corner (weight grid below) and the top row.

This isn't an issue - the top row is duplicated and referenced using inrow[-1] so the maths works out. See comments inline for details.

> I also want to double check what happens when a jpeg has say an odd number width but is still h2v2. IIRC it does something like GL_CLAMP_TO_EDGE and still has a complete macroblock. I think the computations would still work as expected, but I want to double check.

This is correct - the upsampled pixel data in the buffer is clipped to the image size when moving to the frame buffer.

Patch-set: 1
Reviewer: Gerrit User 1247893 <1247893@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>