Update patch set 2
Patch Set 2:

> When we roll this, we can include updated expected result images.

So if I understand correctly, the process for generating the new expected results would be:
1) Make a DEPS roll patch to pull in this CL.
2) Kicks the bots on the DEPS roll.
3) Copy the 'actual' output images from all of the failing JPEG tests and add them to the DEPs roll as the new 'expected' images for the respective tests.

I'm thinking it would be sensible to make a DEPS roll for the change in behaviour (this patch) where we update the expected images and a separate one for the patches containing the SIMD stubs and NEON implementation of h1v2 fancy upsampling.

Confirm if that sounds good to you and I'll go ahead.

Patch-set: 2
Reviewer: Gerrit User 1247893 <1247893@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>