Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'd be delighted to help out as an owner here but could you just clarify the scope of the role given I'm an @arm.com contributor?
> 
> I'm happy to take responsibility for Arm-related things and to help out with platform-agnostic bugs, reviews, etc. My one concern is that we (Arm) don't want to become responsible for the huge amount of Intel assembly code.

Agreed. I don't think it would make sense for you to work on anything Intel-specific.
Not only would that perhaps be a conflict of interest with your employer. It also just doesn't seem like the best allocation of all of our time.

If you are asked to review any Intel-specific things you can reassign to either Leon or I for review. This is true in general as well. If a reviewer is busy they can say so and suggest someone else review.

Patch-set: 1
Reviewer: Gerrit User 1189229 <1189229@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>