Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> This patch is performance neutral for 32-bit - we're replacing a NEON assembly implementation. (Converting to intrinsics has huge benefits in terms of maintainability.)

I know we're moving ahead right now relative to upstream, but eventually we'd like to get this in the upstream repo as well. It seems this patch runs counter to DRC's comments in https://github.com/libjpeg-turbo/libjpeg-turbo/pull/267#issuecomment-410000510. (Correct me if I'm wrong.) How do you plan to convince DRC that this is the right way to go?

Patch-set: 2