Fix duplication of static const data structure which was adding an extra 65k to chrome.dll and chrome_child.dll
While doing some size analysis of chrome binaries I noticed that the jpeg_nbits_table in the libjpeg_turbo codebase was being duplicated.  A unique copy was in the jchuff.obj and jcphuff.obj resulting in an added 65k in .rdata in chrome.dll and chrome_child.dll.  Specifying extern __declspec(selectany) const instead of static const fixes this so only one copy is referenced.

Change-Id: I57ea19dd88951b3ba6b5f978bb53fc5377956c30