Update patch set 1
Patch Set 1:

A few thoughts...

Thanks for doing this. Avoiding duplication of data seems worthwhile. However I wonder if we need to push this further upstream, so that we won't lose the change if we pull the latest version again. Alternately, as we diverge we usually try to list all of the patches that we have added - see README.chromium - although I don't see such a list anywhere.

What about non-Windows platforms? Do they support __declspec(selectany)? If not then we need some #ifdefs. The other way to fix issues like this is to change the array definition from "static const" to "extern const" and put it in a .cc file, and just have a declaration in the header file. This method is potentially more portable, and avoids reparsing the array every time the header file is included.

The git standard is that the subject line should be about 50 characters and the body should generally wrap at around 72 characters. I like this guide:
https://chris.beams.io/posts/git-commit/

Having a shorter subject line is the most important, so that git log --oneline works better.

As for who else, scroggo@chromium.org and cblume@chromium.org are in the OWNERS file.

Note that after you have landed this you will need to "roll-deps" to get Chromium to pull in the latest version of this repo. That's easy enough - let me know if you need advice.

This reminds me that I have a tiny and similar change to libjpeg_turbo, which I just uploaded to here: https://chromium-review.googlesource.com/c/chromium/deps/libjpeg_turbo/+/1684470

Patch-set: 1
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>