Avoid defining arrays in header files
Arrays should not be defined in header files. If they are declared
without static or const then there will be linker errors about
duplicate definitions. If they are declared as static or const then
this will cause the linker to silently allow the duplication definitions
which means that memory is wasted. Not a lot, but it is pure waste.
See this issue for details:

https://github.com/libjpeg-turbo/libjpeg-turbo/issues/114

Unfortunately that issue also explains why this change can't be landed.

Change-Id: Ie3354068f68750633a94bb2b64334198f391a5cf