Update patch set 2
Patch Set 2:

(4 comments)

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > I know we're moving ahead right now relative to upstream, but eventually we'd like to get this in the upstream repo as well. It seems this patch runs counter to DRC's comments in https://github.com/libjpeg-turbo/libjpeg-turbo/pull/267#issuecomment-410000510. (Correct me if I'm wrong.) How do you plan to convince DRC that this is the right way to go?
> > 
> > DRC doesn't want a mixture of NEON assembly code and intrinsics so we've decided to re-write all of the existing NEON assembly code using intrinsics. Once we're finished, there will be no NEON assembly code in libjpeg-turbo.
> 
> Ok, great, thanks! Leaving to cblume to review the assembly.

Sorry about the delay there.

Mostly LGTM.

BTW, I strongly prefer intrinsics over asm. The compiler needs to disable more and more optimizations the closer it gets to the asm. For example, it cannot freely move code around. Intrinsics grant the compiler this insight and allow optimizations to stay. So even with the maintenance improvement aside I am happy with this move. :)



Fun side story: MSVC has a x86 intrinsic __readeflags() which is buggy. I reported this long ago here [1] -- it was confirmed but closed WontFix since no one had noticed before. (Apparently Microsoft Connect has been retired and that link no longer works.)

The core problem is the optimizer could change an ADD instruction (which effects the overflow flag in the eflags) to a LEA dest, [src1+src2] (which does not effect the overflow flag). As a result, the optimizer saw it could move the __readeflags() instruction around since it wasn't waiting on the result of the addition. Of course, I would get incorrect reading that there was no overflow when there had in fact been one. And worse, inserting compiler barriers wasn't enough. It would move the instruction outside the barriers!!

At the end of the day, there is no way to tell the compiler which part of the eflags you intend to use and thus what NOT to optimize away. So __readeflags() is effectively useless. We have to use asm. :(

I hope some day we'll get intrinsics like __readoverflowflag() to tell the compiler what not to optimize away.

[1] http://connect.microsoft.com/VisualStudio/feedback/details/691456/-readeflags-intrinsic-can-be-reordered-by-the-compiler

Patch-set: 2
Reviewer: Gerrit User 1189229 <1189229@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>