Update patch set 4
Patch Set 4:

Thank you for finding and fixing this.
And congrats on your first Chromium PR!

The README.chromium [1] does indeed have the list of changes from upstream. It is the second bulleted list, starting with "Also, we applied the following changes which are not merged to upstream:"

You'll want to edit that file and add a description of the change. That should be done as part of this PR. (BTW we typically call these CLs for Change List. And each uploaded update is a PS for Patch Set).


Also, we will handle upstreaming this to libjpeg-turbo if you want. Upstreaming is in a tricky spot right now [2]. Maybe this would count as a bug fix (since the duplication is certainly not intentional). But maybe it counts as a feature (since it is improving binary size). *shrugs*

Our goal is to have all of our local changes upstreamed. That doesn't always work out though because of things like [2].

If you want to create a parallel PR for upstream [3] you're more than welcome. If not, I can try to preserve commit history across repos to keep the change tagged with your name so you get proper credit when we upstream it.


Other than that, I like it. :D Just update the README.chromium and we're good to go.


[1] https://cs.chromium.org/chromium/src/third_party/libjpeg_turbo/README.chromium
[2] https://libjpeg-turbo.org/About/Help
[3] https://github.com/libjpeg-turbo/libjpeg-turbo

Patch-set: 4
Reviewer: Gerrit User 1189229 <1189229@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>