Update patch set 1
Patch Set 1:

> Would you mind posting the perf & binary size differences?

MonochromePublic.apk (as far as I know, the one the binary-size bot tests) increased by 4.1KB. I could be building this the wrong way though... I suppose the bot will tell us when we do the DEPS roll. If the bot goes red on the CQ dry-run and the increase is deemed acceptable we can add the "binary-size: <justification>" thing to the DEPS roll commit message to silence it.

Performance justification:

I've been using the 'Kevin' Chromebook (A72 and A53 cores) as my benchmark device to validate my libjpeg-turbo NEON patches. I've been wanting to test on newer hardware for a while and last week I got my hands on a Pixel 3 (A75 and A55 cores); I got a bit of a surprise re-benchmarking the color conversion patch [1].

On the A72 and A53 cores the patch is fine - performance neutral on 32-bit for both cores and a modest benefit on 64-bit. However, benchmarking on the Pixel 3 I saw a 3-10% regression for 32-bit on the A55 core. Increasing the compiler optimization level in the libjpeg-turbo BUILD.gn file to -O3 using "optimize_speed" more than cures the issue - hence this CL. I'll attach the spreadsheet with the data to the tracker bug. 

[1] https://chromium-review.googlesource.com/c/chromium/deps/libjpeg_turbo/+/1717294/4

P.S. did my email (sent of Tuesday) with this information + the spreadsheet bounce?

Patch-set: 1
Reviewer: Gerrit User 1247893 <1247893@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>