Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> It seems like these two files are the only ones that reference jpeg_nbits_table.inc. They are now using the table that is defined in C++ (rather than the .inc which defines it in asm).
> 
> Since this is the case, we might as well delete jpeg_nbits_table.inc. If we do that, we should update the part of README.chromium which says which files have been deleted.
> 
> 
> As a separate thought, I feel like the difference between CPU speed and memory latency makes these tables (which are easily reproducible with CPU-instructions) kinda useless. The table is:
> 1x0
> 1x1
> 2x2
> 4x3
> 8x4
> 16x5...etc
> 
> I think the whole table can be replaced with finding the position of the highest bit set (and special casing 0). Said another way, CLZ - TOTAL_BITS. (Again, special casing 0)
> 
> Removing the tables entirely would also free up that cache line.
> 
> What do you you about possibly removing the table entirely?
I removed the .inc file.  

I did see that the lookup table was easily to compute on the fly, but the comments in the code suggested it needed to be there as-is.  See the comments in jchuff.c and jcphuff.c.  Specifically about the usage of USE_CLZ_INTRINSIC which bypasses the lookup table entirely.

Patch-set: 5
Reviewer: Gerrit User 1354609 <1354609@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>