Mark jpeg_nbits_table as external in asm files to reduce disk impact
A recent change was made to mitigate the duplication of the
jpeg_nbits_table array in libjpeg_turbo that resulted in 64k in disk
impact for each module it was included in.  That change did not mitigate
the usage of jpeg_nbits_table in the assembly files in that library,
thus the duplication remains.  This change removes the definition of the
array in those files and replaces it with a declaration of an external
reference.

Previous change: http://crrev.com/c/1685791

Bug: http://crbug.com/992072
Change-Id: I6c7f1b3a665f36b81eacd32050f69b4dca932b1e