Update patch set 6
Patch Set 6:

> I am not able to find the unit tests for libjpeg-turbo on their github page.  

See [1] for instructions on how to build and test the upstream libjpeg-turbo - the section on regression testing is right at the bottom. I don't think we mind too much which x86(_64) target you build and test for since they all use the same code.

> I don't have permission to submit the change.  In the past, Chris B had to do that on my behalf. 

Sure, that's no problem - I'm asking that the change is rebased prior to submission so we don't end up with diamonds and merge commits in the tree. If you do 'git pull --rebase' in src/third_party/libjpeg_turbo (in your Chromium checkout) and then re-upload the change we should be golden. Alternatively, you should be able to do this using the "REBASE" button in the top right corner - I don't think you need any special permissions to use that. 

[1] https://github.com/libjpeg-turbo/libjpeg-turbo/blob/master/BUILDING.md

Patch-set: 6