Update patch set 1
Patch Set 1:

PTAL.

Median performance gains:
* 64-bit: 6.7% (big); 4.6% (little).
* 32-bit: -0.2% (big); 1.4% (little).

I have attached a spreadsheet containing the benchmark data to the tracker bug.

The very minor median regression for 32-bit on the big core is not a problem: we have already submitted patches which benefit h2v2, h1v2 and h2v1 sampled images. For h1v1 images, the median performance gain (brought by this patch) for 32-bit on the big core is 1.6%. (This data and graphs are available in the spreadsheet.)

This patch will likely trigger a regression on the Chromium Perf bots (again see spreadsheet) - which should be marked as "will not fix" because the bots are testing only a single image on 32-bit (while the data for the whole Google JPEG corpus is overwhelmingly positive.)

Patch-set: 1
Reviewer: Gerrit User 1189229 <1189229@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1000164 <1000164@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>