Update patch set 1
Patch Set 1:

It would probably be good to have a design doc to describe what changes are being made and why. Also, we should list any diffs from upstream in README.chromium.

That said, it might be worth it to see if upstream would be amenable to some kind of refactoring for this. (Not sure - this might be considered a build system enhancement, according to https://github.com/libjpeg-turbo/libjpeg-turbo/blob/master/.github/CONTRIBUTING.md)

Patch-set: 1