Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> > Patch Set 2:
> > 
> > Hopefully you both have a link to the design doc I made.
> 
> Yes, thanks! Would it be possible to make it publicly viewable and link to it from the bug/CL?

Doc is now public and the link added as a comment on crbug 993876.
 
> > As for pushing this to the upstream project, I'd be highly surprised if this got accepted - though I'd be happy to give it a go after we get this working here.
> 
> Yeah, I guess the main reason we need it is to otherwise diverge from upstream, so maybe you're right. I'd at least like to do our best to make this least disruptive to future changes. And even better, if new tests are added, we'd want our test infrastructure to run them without us knowing they were added and having to update our code. Which I think argues *against* splitting up this test - if new tests are added upstream, we run this as one test, and there aren't any merge conflicts (or symbol collisions...), we should pick up the new tests automatically. (Maybe that's all too much to hope for.) Do you expect the upside to splitting up the tests to be worth it? (The commit message doesn't go into detail on why you split them up, but the design doc says "To increase test visibility and reduce debugging effort". Do you expect we'll break the tests and need to spend time debugging?)

I do think this is worth splitting up. In my experience of rewriting the NEON code paths for this library and trying many tricks along the way (that I thought I could get away with according to the libjpeg API) this test is the one that annoyed me most. As it currently stands, it will - if you've made a mistake - segfault and all you know is that "something... somewhere... broke" among the myriad of test cases and you have to start commenting out test cases in main() to narrow things down. 

Even if we had a different test case for every different combination of top-level command line switches, each test is still running through 14 or 21 configurations which, IMO, is way too many.

To be honest, if more tests were going to be added upstream, I'd expect them to be for more cjpeg and djpeg cases - which, because they're very granular and differentiated by the command line switches, we'd have to write new GTests manually for anyway. Sadly, I don't think we're going to be getting much free here when updating from upstream.

Patch-set: 2