Update patch set 1
Patch Set 1:

This patch gets a bit interesting because the tjbench tests require the source image to be copied (and renamed) into the current execution directory. This introduces a dependency on 'base' - which I'm not entirely happy about as it means Chromium's libjpeg-turbo library isn't independent of chromium/src.

When all of this is done we want to use Chromium's libjpeg-turbo in AOSP and ideally we would want to run these tests there too. This gets difficult with the tests dependent on 'base'. Adding Elliott Hughes to get his thoughts / a way around this in AOSP.

(File and path manipulation operations will be required in subsequent tests to be added for cjpeg, djpeg and jpegtran so this isn't a problem applicable to only this set of tests.)

Patch-set: 1
Reviewer: Gerrit User 1189229 <1189229@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1000164 <1000164@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1004789 <1004789@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>