Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > > The filesystem stuff seems more tricky but let's have a go...
> > > 
> > > well,
> > > ```
> > > 
> > >     base::FilePath resource_path;
> > >     ASSERT_TRUE(base::PathService::Get(base::DIR_SOURCE_ROOT, &resource_path));
> > >     resource_path = resource_path.AppendASCII("third_party");
> > >     resource_path = resource_path.AppendASCII("libjpeg_turbo");
> > >     resource_path = resource_path.AppendASCII("testimages");
> > >     resource_path = resource_path.AppendASCII("testorig.ppm");
> > >     ASSERT_TRUE(base::PathExists(resource_path));
> > > ```
> > > 
> > > was never going to work on Android anyway, so i doubt it makes much difference to the need for an #if...
> > 
> > Umm, just to be clear - it is possible for tests to reference resources in the AOSP source tree? If not, AOSP will just have to trust that Chromium has done the necessary testing for libjpeg-turbo and not run the tests there too.
> 
> sort of: you can copy data files to the directory containing the test.

Can you point me at an example in AOSP please? Chromium's GN build system allows me to specify a data dependency on the testimage directory which it will then package properly for whatever platform the tests are going to run on. I don't know anything about AOSP's Soong - or what the right way to do test resource referencing is.

If you like, we can get this landed here as is, land the rest of the NEON intrinsics patches so we get BTI support and worry about actually running the tests in AOSP's infrastructure at a later date? (Though personally, if we can do this right now, I'd prefer that.)

Patch-set: 2