Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> FWIW, I'm comfortable landing this in Chromium.
> I see the discussion about the weird file manipulation and complication for Android. Whatever you all think is best there is fine with me.
> 
> I do think I prefer the std::filesystem over base:: stuff here. It gets into a bit of gray area though. That is C++17, and Chromium is still C++14.

I think it's certainly better if we don't have to depend on base:: - faster build times and we're not tied to a Chromium checkout.

Is there a way to set -std=c++17 for only this testing target? (I've been experimenting with the big hammer of editing //build/config/compiler/BUILD.gn to set -std=c++17 and add the linker flag -lstdc++fs but can't get it to build...)

Patch-set: 2