Specify 16-byte alignment for assembly functions
A previous commit that implemented h2v2 downsampling using Arm NEON
intrinsics removed ".balign 16" along with the NEON assembly
function. This function alignment was the only one of its type in
jsimd_neon.S and seems to have 'accidentally' ensured that all of the
other functions declared in the same file also have that alignment.

Removing the 16-byte alignment specification meant that the Mac arm64
build of Chromium broke - as the remaining assembly functions did not
have the right alignment.

This commit specifies 16-byte alignment for all of the NEON assembly
functions in jsimd_neon.S

Bug: 922430
Change-Id: Ie827cf2166a6ac5e2b69db3d2140ed91ab2798e3